backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.7
eventlet==0.36.1
python-dotenv==1.0.1
redis==5.0.8
SQLAlchemy==2.0.36
PyMySQL==1.1.1
